philo_one: Almost done?

+++ philo_one/utils/libft.h
@@ -6,7 +6,7 @@
 /*   By: mli <marvin@42.fr>                         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/09 12:34:27 by mli               #+#    #+#             */
-/*   Updated: 2020/09/18 14:18:57 by mli              ###   ########.fr       */
+/*   Updated: 2020/09/21 15:14:15 by mli              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -88,8 +88,9 @@ t_list			*ft_lstmap(t_list *lst, void *(*f)(void *),
 
 void			ft_free(void **ptr);
 int				ft_atoi_ptr(const char *str, int *i);
-int				ft_nbrlen_base(long int nb, int base_len);
-int				ft_nbrlen(long int nb);
+int				ft_nbrlen_base(long unsigned int nb,
+		unsigned long int base_len);
+int				ft_nbrlen(long unsigned int nb);
 char			*ft_itoa_lbase(const char *base, long int nbr);
 
 double			ft_abs_lf(double nbr);
